refactor: move ZK ElGamal name map to a pure util
Addresses review feedback: keep the utils -> components layering clean by
extracting the instruction name lookup into app/utils/zk-elgamal-proof.ts,
imported by both the detail card and instruction.ts. Also add a test for
the empty instruction data fallback.
